:recycle: refactor: rework tree architecture of stories and externalise argsType
