+++ index.html
@@ -17,7 +17,7 @@
     <script>
       const test = () => {
         console.log("yo");
-      }
+      };
     </script>
   </head>
   <body>
