+++ src/stories/Accordion/accordion.stories.ts
@@ -1,10 +1,10 @@
-import { StoryObj } from "@storybook/web-components";
+import { StoryObj, Meta } from "@storybook/web-components";
 import { html } from "lit";
 
 import { CandyAccordionProps } from "../../components/Accordion";
 import "../../components/Accordion";
 import "../../components/Alert";
-import { meta } from "./accordion";
+import { meta } from "./AccordionMeta";
 
 type Story = StoryObj<CandyAccordionProps>;
 
@@ -21,9 +21,8 @@ const renderAccordion = (args: CandyAccordionProps) => html`<candy-accordion
 export default {
   ...meta,
   title: "Components/Accordion",
-  component: "candy-accordion",
   render: renderAccordion,
-}
+} as Meta<CandyAccordionProps>;
 
 export const Accordion: Story = {
   args: {
