+++ src/stories/Badge/Badge.stories.ts
@@ -1,8 +1,9 @@
 import { StoryObj, Meta } from "@storybook/web-components";
 import { html } from "lit";
 
-import { CandyBadgeProps } from "../components/Badge";
-import "../components/Badge";
+import { CandyBadgeProps } from "../../components/Badge";
+import "../../components/Badge";
+import { meta } from "./BadgeMeta";
 
 type Story = StoryObj<CandyBadgeProps>;
 
@@ -10,21 +11,8 @@ const renderBadge = (args: CandyBadgeProps) =>
   html`<candy-badge label=${args.label}></candy-badge>`;
 
 export default {
+  ...meta,
   title: "Components/Badge",
-  component: "candy-badge",
-  argTypes: {
-    label: {
-      description: "Text displayed inside the badge",
-      table: {
-        type: {
-          summary: "string",
-        },
-        defaultValue: {
-          summary: "Badge",
-        },
-      },
-    },
-  },
   render: renderBadge,
 } as Meta<CandyBadgeProps>;
