+++ src/stories/Badge/BadgeMeta.ts
@@ -1,0 +1,16 @@
+export const meta = {
+  component: "candy-badge",
+  argTypes: {
+    label: {
+      description: "Text displayed inside the badge",
+      table: {
+        type: {
+          summary: "string",
+        },
+        defaultValue: {
+          summary: "Badge",
+        },
+      },
+    },
+  },
+};
