+++ src/stories/Checkbox/CheckboxMeta.ts
@@ -1,20 +1,4 @@
-import { StoryObj, Meta } from "@storybook/web-components";
-import { html } from "lit";
-
-import { CandyCheckboxProps } from "../components/Checkbox";
-import "../components/Checkbox";
-
-type Story = StoryObj<CandyCheckboxProps>;
-
-const renderCheckbox = (args: CandyCheckboxProps) =>
-  html`<candy-checkbox
-    ?checked=${args.checked}
-    label=${args.label}
-    description=${args.description}
-  ></candy-checkbox>`;
-
-export default {
-  title: "Components/Checkbox",
+export const meta = {
   component: "candy-checkbox",
   argTypes: {
     checked: {
@@ -50,14 +34,5 @@ export default {
         },
       },
     },
-  },
-  render: renderCheckbox,
-} as Meta<CandyCheckboxProps>;
-
-export const Checkbox: Story = {
-  args: {
-    checked: false,
-    label: "Checkbox",
-    description: "This is a description",
   },
 };
