+++ src/stories/Sidebar/Element/Element.stories.ts
@@ -3,6 +3,7 @@ import { html } from "lit";
 
 import { CandySidebarElementProps } from "../../../components/Sidebar/Element";
 import "../../../components/Sidebar/Element";
+import { meta } from "./ElementMeta";
 
 type Story = StoryObj<CandySidebarElementProps>;
 
@@ -13,46 +14,9 @@ const renderElement = (args: CandySidebarElementProps) => html`<candy-sidebar-el
 ></candy-sidebar-element>`;
 
 export default {
+  ...meta,
   title: "Components/Sidebar/Element",
-  component: "candy-sidebar-element",
   render: renderElement,
-  argTypes: {
-    label: {
-      description: "Label of the element",
-      table: {
-        type: {
-          summary: "string",
-        },
-        defaultValue: {
-          summary: "Home",
-        },
-      },
-    },
-    active: {
-      name: "?active",
-      description: "Is the item the selected one",
-      table: {
-        type: {
-          summary: "boolean",
-        },
-        defaultValue: {
-          summary: "False",
-        },
-      },
-    },
-    disabled: {
-      name: "?disabled",
-      description: "Is the item disabled",
-      table: {
-        type: {
-          summary: "boolean",
-        },
-        defaultValue: {
-          summary: "False",
-        },
-      },
-    },
-  },
 } as Meta<CandySidebarElementProps>;
 
 export const Element: Story = {
