+++ src/stories/Sidebar/Element/ElementMeta.ts
@@ -1,0 +1,40 @@
+export const meta = {
+  component: "candy-sidebar-element",
+  argTypes: {
+    label: {
+      description: "Label of the element",
+      table: {
+        type: {
+          summary: "string",
+        },
+        defaultValue: {
+          summary: "Home",
+        },
+      },
+    },
+    active: {
+      name: "?active",
+      description: "Is the item the selected one",
+      table: {
+        type: {
+          summary: "boolean",
+        },
+        defaultValue: {
+          summary: "False",
+        },
+      },
+    },
+    disabled: {
+      name: "?disabled",
+      description: "Is the item disabled",
+      table: {
+        type: {
+          summary: "boolean",
+        },
+        defaultValue: {
+          summary: "False",
+        },
+      },
+    },
+  },
+};
