+++ src/stories/Sidebar/SidebarUseCase.stories.ts
@@ -9,10 +9,10 @@ import { Sidebar } from "./Sidebar.stories";
 type Story = StoryObj<CandySidebarProps & CandySideBarControl>;
 
 const renderUseCaseSideBar = (args: CandySidebarProps & CandySideBarControl) => {
-  return html`<div style="display: flex; background-color: #F8F9FA; border-radius: 8px; min-height: 350px">
-    <div>
-      ${renderSideBar(args)}
-    </div>
+  return html`<div
+    style="display: flex; background-color: #F8F9FA; border-radius: 8px; min-height: 350px"
+  >
+    <div>${renderSideBar(args)}</div>
     <div
       style="
         background-color: #D2D4D6;
@@ -22,10 +22,11 @@ const renderUseCaseSideBar = (args: CandySidebarProps & CandySideBarControl) => 
         display: flex;
         justify-content: center;
         align-items: center;
-    ">
+    "
+    >
       <p>CONTENT</p>
     </div>
-  </div>`
+  </div>`;
 };
 
 export default {
