+++ src/stories/SidebarDoc/Element/DocSidebarTitleMeta.ts
@@ -1,0 +1,28 @@
+export const meta = {
+  component: "candy-sidebar-doc-title",
+  argTypes: {
+    label: {
+      description: "Label of the element",
+      table: {
+        type: {
+          summary: "string",
+        },
+        defaultValue: {
+          summary: "Home",
+        },
+      },
+    },
+    active: {
+      name: "?active",
+      description: "Is the item the selected one",
+      table: {
+        type: {
+          summary: "boolean",
+        },
+        defaultValue: {
+          summary: "False",
+        },
+      },
+    },
+  },
+};
