+++ src/stories/Spinner/SpinnerMeta.ts
@@ -1,0 +1,20 @@
+import { SpinnerSizes } from "../../components/Spinner";
+
+export const meta = {
+  component: "candy-spinner",
+  argTypes: {
+    size: {
+      description: "Size of the spinner",
+      options: Object.values(SpinnerSizes),
+      control: { type: "radio" },
+      table: {
+        type: {
+          summary: "string",
+        },
+        defaultValue: {
+          summary: SpinnerSizes.md,
+        },
+      },
+    },
+  },
+};
