+++ src/stories/Switch/Switch.stories.ts
@@ -1,8 +1,9 @@
 import { StoryObj, Meta } from "@storybook/web-components";
 import { html } from "lit";
 
-import { CandySwitchProps } from "../components/Switch";
-import "../components/Switch";
+import { CandySwitchProps } from "../../components/Switch";
+import "../../components/Switch";
+import { meta } from "./SwitchMeta";
 
 type Story = StoryObj<CandySwitchProps>;
 
@@ -10,22 +11,9 @@ const renderSwitch = (args: CandySwitchProps) =>
   html`<candy-switch ?activated=${args.activated}></candy-switch>`;
 
 export default {
+  ...meta,
   title: "Components/Switch",
-  component: "candy-switch",
   render: renderSwitch,
-  argTypes: {
-    activated: {
-      description: "Boolean value of the switch",
-      table: {
-        type: {
-          summary: "boolean",
-        },
-        defaultValue: {
-          summary: "False",
-        },
-      },
-    },
-  },
 } as Meta<CandySwitchProps>;
 
 export const Switch: Story = {
