+++ src/stories/Switch/SwitchMeta.ts
@@ -1,0 +1,16 @@
+export const meta = {
+  component: "candy-switch",
+  argTypes: {
+    activated: {
+      description: "Boolean value of the switch",
+      table: {
+        type: {
+          summary: "boolean",
+        },
+        defaultValue: {
+          summary: "False",
+        },
+      },
+    },
+  },
+};
